requirements.txt
streamlit
Pillow
PyMuPDF
numpy
